Return error message if file does not exist.

+++ bibos_global_scripts/remove_desktop_shortcut.sh
@@ -8,7 +8,16 @@ then
 fi
 
 FILENAME=$1
+FILEPATH="/home/.skjult/Desktop/$FILENAME"
 
-rm -f "/home/.skjult/Desktop/$FILENAME"
+if [ -f $FILEPATH ]
+then
+    rm -f "/home/.skjult/Desktop/$FILENAME"
+else
+    echo "File does not exist"
+    exit -1
+fi
+
+
 
 exit 0
